removed shadow dom stuff
